requirements.txt
yfinance>=0.2.38
statsmodels>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
requests>=2.31.0
streamlit>=1.32.0
plotly>=5.20.0
schwab-py>=1.0.0
supabase>=2.0.0
